refactor: entities 폴더에 슬라이스 추가, 기존 파일을 세그먼트에서 관리하도록 이동
